Cognito credentials through proxy test
